Port changes to use size_t for memory allocations to 64bit machine.
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%403068

+++ mpqc/src/lib/chemistry/qc/intv3/init2e.cc
@@ -56,7 +56,7 @@ fail()
  * integrals are stored.
  */
 double *
-Int2eV3::int_initialize_erep(int storage, int order,
+Int2eV3::int_initialize_erep(size_t storage, int order,
                              const Ref<GaussianBasisSet> &cs1,
                              const Ref<GaussianBasisSet> &cs2,
                              const Ref<GaussianBasisSet> &cs3,
